OnboardingViewの画面遷移メソッドをモデルに分離した
